+++ matjalalzz/src/main/java/shop/matjalalzz/mock/MockUser.java
@@ -1,4 +1,4 @@
-package shop.matjalalzz.domain.mock;
+package shop.matjalalzz.mock;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -15,10 +15,12 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 @AllArgsConstructor
-public class MockParty {
+public class MockUser {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    private String nickname;
 
 }
