+++ matjalalzz/src/main/java/shop/matjalalzz/party/mock/dao/MockShopRepository.java
@@ -1,8 +1,8 @@
 package shop.matjalalzz.party.mock.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shop.matjalalzz.party.mock.entity.MockShop;
+import shop.matjalalzz.party.mock.entity.MockShop2;
 
-public interface MockShopRepository extends JpaRepository<MockShop, Long> {
+public interface MockShopRepository extends JpaRepository<MockShop2, Long> {
 
 }
