+++ matjalalzz/src/main/java/shop/matjalalzz/review/dto/entity/ReviewCreateRequest.java
@@ -1,4 +1,4 @@
-package shop.matjalalzz.domain.review.dto.entity;
+package shop.matjalalzz.review.dto.entity;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
