+++ matjalalzz/src/main/java/shop/matjalalzz/review/entity/Review.java
@@ -1,6 +1,5 @@
-package shop.matjalalzz.domain.comment.entity;
+package shop.matjalalzz.review.entity;
 
-import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -10,25 +9,28 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
+import jakarta.validation.constraints.Digits;
 import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shop.matjalalzz.domain.mock.MockParty;
-import shop.matjalalzz.domain.mock.MockUser;
-import shop.matjalalzz.global.unit.BaseEntity;
+import shop.matjalalzz.global.common.BaseEntity;
+import shop.matjalalzz.mock.MockImage;
+import shop.matjalalzz.mock.MockReservation;
+import shop.matjalalzz.mock.MockShop;
+import shop.matjalalzz.mock.MockUser;
 
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder
 @AllArgsConstructor
-public class Comment extends BaseEntity {
+public class Review extends BaseEntity {
 
     @Id
-    @Column(name = "comment_id")
+    @Column(name = "review_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
@@ -37,21 +39,26 @@ public class Comment extends BaseEntity {
     private String content;
 
     @Column(nullable = false)
+    @Digits(integer = 1, fraction = 2)
+    private Double rating = 0D;
+
+    @Column(nullable = false)
     private boolean deleted = false;
-
-    @ManyToOne
-    private Comment parent;
-
-    @OneToMany(cascade = CascadeType.REMOVE, orphanRemoval = true)
-    private List<Comment> children;
 
     @ManyToOne(optional = false)
     @JoinColumn(nullable = false)
-    private MockParty party;
+    private MockShop shop;
+
+    @ManyToOne(optional = false)
+    @JoinColumn(nullable = false)
+    private MockReservation reservation;
 
     @ManyToOne(optional = false)
     @JoinColumn(nullable = false)
     private MockUser writer;
+
+    @OneToMany(mappedBy = "review")
+    private List<MockImage> images;
 
 
 }
